Clip weights so they don't go below zero
