+++ auramask/callbacks/__init__.py
@@ -4,3 +4,4 @@ from auramask.callbacks.metriclogger import AuramaskWandbMetrics
 from auramask.callbacks.stop_on_nan import AuramaskStopOnNaN
 from auramask.callbacks.evaluate import AuramaskCallback
 from auramask.callbacks.checkpoint import AuramaskCheckpoint
+from auramask.callbacks.softadapt import AdaptiveLossCallback
